lead_management/serializers/status_serializer.py
from rest_framework import serializers

from lead_management.models import Status


class StatusSerializer(serializers.ModelSerializer):
    class Meta:
        model = Status
        fields = '__all__'
