requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backports.weakref==1.0.post1
beautifulsoup4==4.12.3
billiard==4.2.0
boto3==1.34.62
botocore==1.34.62
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==42.0.5
cssselect2==0.7.0
decorator==5.1.1
distro==1.9.0
Django==4.1.5
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-requests-logger==0.1.7
django-smtp-ssl==1.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.6.1
drf-yasg==1.21.5
et-xmlfile==1.1.0
eventlet==0.35.2
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
fuzzywuzzy==0.18.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httpx==0.27.0
idna==3.6
inflection==0.5.1
ipython==8.18.1
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.5
Levenshtein==0.25.1
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
multidict==6.0.5
numpy==1.26.4
openai==1.14.0
openpyxl==3.1.0
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==24.0
pandas==1.5.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
pip==23.0.1
prompt-toolkit==3.0.43
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.18.0
pyHanko==0.23.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.1.0
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
rapidfuzz==3.9.0
regex==2023.12.25
reportlab==4.0.9
requests==2.31.0
rest_framework_roles==1.0.4
rollbar==1.0.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
scipy==1.12.0
selenium==4.18.1
setuptools==58.1.0
six==1.16.0
slackclient==2.9.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.66.2
traitlets==5.14.3
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
undetected-chromedriver==3.5.5
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==12.0
whitenoise==6.6.0
wsproto==1.2.0
xhtml2pdf==0.2.15
yarl==1.9.4
